He mejorado bastantes cosas

+++ resources/js/admin/CustomButtonComponent.js
@@ -1,4 +1,5 @@
 import { GridManager } from "./GridManager";
+import Swal from "sweetalert2";
 export class CustomButtonComponent {
     eGui;
     eButton;
@@ -6,19 +7,36 @@ export class CustomButtonComponent {
 
     init(param) {
         this.eGui = document.createElement('div');
-        const eButton = document.createElement('button');
-        eButton.className = 'btn btn-danger';
-        eButton.textContent = 'Eliminar';
+        const eButton = document.createElement('div');
+        eButton.className = 'delete-btn';
+        eButton.innerHTML = '<i class="fa fa-trash-o" aria-hidden="true"></i>';
         this.eventListener = () => {
-            GridManager.gridApi.applyTransaction({ remove: [{ id: param.data.id }] });
-            let result = GridManager.delete(`${GridManager.url}/${param.data.id}`);
-            console.log(result);
+            Swal.fire({
+                title: "Estás seguro?",
+                text: "No podrás revertir esto!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Eliminar"
+            }).then(async (result) => {
+                if (result.isConfirmed) {
+                    let result = await GridManager.delete(`${GridManager.url}/${param.data.id}`);
+                    console.log(result);
 
-
+                    if (result.status === 200) {
+                        GridManager.gridApi.applyTransaction({ remove: [{ id: param.data.id }] });
+                        Swal.fire({
+                            title: "Eliminado!",
+                            text: result.message,
+                            icon: "success"
+                        });
+                    }
+                }
+            });
         }
         eButton.addEventListener('click', this.eventListener);
         this.eGui.appendChild(eButton);
-        console.log("hola");
 
     }
